Switched to new hexabus protocol version

+++ hostsoftware/libjhexabus/src/de/fraunhofer/itwm/hexabus/HexabusEndpointQueryPacket.java
@@ -1,7 +1,7 @@
 package de.fraunhofer.itwm.hexabus;
 public class HexabusEndpointQueryPacket extends HexabusQueryPacket {
-	private int eid;
-	public HexabusEndpointQueryPacket(int eid) throws Hexabus.HexabusException {
+	private long eid;
+	public HexabusEndpointQueryPacket(long eid) throws Hexabus.HexabusException {
 		super(eid);
 		this.packetType = Hexabus.PacketType.EPQUERY;
 	}
